Move DSL to its own package
